Move main.Config to a subpackage
So that we can use it in the integration tests.

See  https://github.com/golang/go/issues/5198
